backend/requirements.txt
fastapi
uvicorn
websockets
twilio
motor
google-generativeai
elevenlabs
python-dotenv
